+++ backend/src/test/java/com/app/backend/domain/group/repository/GroupRepositoryTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.app.backend.domain.group.entity.Group;
 import com.app.backend.domain.group.entity.RecruitStatus;
-import com.app.backend.global.supporter.SpringBootTestSupporter;
+import com.app.backend.domain.group.supporter.SpringBootTestSupporter;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
